work with dynamic arrays
