+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "files.associations": {
+        "dynamic_arrays.h": "c",
+        "stdlib.h": "c",
+        "dynamic_head.h": "c"
+    }
+}
