+++ 2_spusk_andreevskii/dynamic_head.h
@@ -1,0 +1,12 @@
+#ifndef DYNAMIC_ARRAYS_H
+#define DYNAMIC_ARRAYS_H
+
+int sum_elements_of_array(int array[], int size);
+
+int find_minimal_elem_of_array(int array[], int size);
+
+int find_maximal_elem_of_array(int array[], int size);
+
+int reverse_array_of_integers(int array[], int size);
+
+#endif
