--- 5_arrays/2_dynamic_arrays/global.c
@@ -1,2 +1,0 @@
-int *integers_array;
-int size;
