--- 5_arrays/2_dynamic_arrays/global.h
@@ -1,2 +1,0 @@
-extern int *integers_array;
-extern int size;
